+++ package.json
@@ -25,7 +25,7 @@
     "eslint-config-airbnb-base": "^14.2.0",
     "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-jest": "^24.1.0",
-    "htmlhint": "^0.9.9",
+    "htmlhint": "^0.16.0",
     "jest": "^27.0.1",
     "regenerator-runtime": "^0.13.1",
     "serve": "^12.0.0",
@@ -34,5 +34,8 @@
   },
   "engines": {
     "node": ">=10"
+  },
+  "dependencies": {
+    "firebase": "^9.2.0"
   }
 }
