+++ src/components/wallPage.js
@@ -12,7 +12,16 @@ export const wallPage = () => {
   const logo = document.createElement("logoImage");
   logo.className = "logoImage";
   logo.src = "images/logo-petsbook.png";
+  header.appenChild(logo);
+
+  const home = document.createElement('section');
+  home.className = 'homeSection';
+  const post =`
+  <div class="post-box">
+  <form id="postBox" class="container-post">
+  </form>
+  </div>`
+  
 
   return wallContainer + introPage;
 };*/
-
