+++ src/index.html
@@ -1,14 +1,28 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>Petsbook</title>
-</head>
-<body>
-  <div id="root"></div>
-  <script type="module" src="index.js"></script>
-  <script type="module" src="main.js"></script>
-</body>
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
+    <title>Petsbook</title>
+  </head>
+  <body>
+    <!--<div class="input-container">
+    <input type="email" id="mailRegister" placeholder="correo@example.com">
+    <input type="password" id="passwordRegister" placeholder="contraseña">
+  </div>
+  <div class="boton-register">
+    <button id="btnRegister" class="btn-Register">Iniciar Sesión</button>
+  </div>
+       <div class="input-container">
+         <input type="email" id="mailLogin" placeholder="correo@example.com">
+         <input type="password" id="passwordLogin" placeholder="contraseña">
+       </div>
+       <div class="btnLogin">
+         <button id="btnLogin" class="btnLogin">Iniciar Sesión</button>
+       </div>-->
+    <div id="root"></div>
+    <script type="module" src="paths.js"></script>
+    <script type="module" src="main.js"></script>
+  </body>
 </html>
