+++ src/main.js
@@ -1,15 +1,9 @@
-// Este es el punto de entrada de tu aplicacion
+// import { routes } from "./paths.js";
 
-// import { myFunction } from './lib/index.js';
+// window.addEventListener("load", () => {
+//   routes(window.location.hash);
+// });
 
-// myFunction();
-
-import { routes } from "./index.js";
-
-window.addEventListener("load", () => {
-  routes(window.location.hash);
-});
-
-window.addEventListener("hashchange", () => {
-  routes(window.location.hash);
-});
+// window.addEventListener("hashchange", () => {
+//   routes(window.location.hash);
+// });
